+++ resources/lang/en/utils.php
@@ -43,5 +43,6 @@ return [
     'impostor_active' => '¡There is an impostor in the class!',
 
     'get_challenge' => 'Congratulations! You\'ve passed the challenge!',
+    'by_group' => 'By group',
 
 ];