+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 ### Changed
 - Change edit for save when editing objects.
+- Fix gold update in battles.
 
 ## v1.64 Beta (09/07/2021)
 ### Changed